libeemd: added a reference to original repository and a patch comment
A head tag for the original bitbucket master branch was added, and
also a doi tag is added to point the original paper.

Based on the review comment, the reason why a local patch was
required was mentioned.  The patch fixes the Makefile build option
to use the -dynamiclib option instead of the -shared option when
making a macOS dynamic link library and also fixes the dynamic
link library suffix name to follow the name convention used in
macOS. Since the original Makefile does not support multi-platform
configuration, we handle this with a local patch until the
original author switches to use autoconf or some other flexible
build environment adaptation tools.